Prevent a Layer from being a child of multiple GroupLayers

+++ core/src/forplay/core/GroupLayerImpl.java
@@ -1,5 +1,18 @@
-//
-// $Id$
+/**
+ * Copyright 2011 The ForPlay Authors
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 
 package forplay.core;
 
@@ -21,13 +34,21 @@ public class GroupLayerImpl<L extends AbstractLayer>
   }
 
   public void add(GroupLayer self, int index, L child) {
+    // remove the child from any existing parent, preventing multiple parents
+    if (child.parent() != null) {
+      child.parent().remove(child);
+    }
     children.add(index, child);
     child.setParent(self);
     child.onAdd();
   }
 
   public void remove(GroupLayer self, L child) {
-    children.remove(child);
+    boolean wasRemoved = children.remove(child);
+    if (!wasRemoved) {
+      throw new UnsupportedOperationException(
+          "Could not remove Layer because it is not a child of the GroupLayer");
+    }
     child.onRemove();
     child.setParent(null);
   }
